pyproject.toml
[project]
name = "agent-learn"
version = "0.1.0"
description = "ai agent examples"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix-otel>=0.9.2",
    "dspy>=2.6.21",
    "fast-flights>=2.2",
    "helium>=5.1.1",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.6",
    "matplotlib>=3.10.1",
    "openinference-instrumentation-dspy>=0.1.21",
    "openinference-instrumentation-litellm>=0.1.19",
    "pillow>=11.2.1",
    "pydub>=0.25.1",
    "python-magic>=0.4.27",
    "selenium>=4.31.0",
    "smolagents[litellm,openai,telemetry]>=1.14.0",
    "yfinance>=0.2.57",
]
